chore(PR): move mutable clock, base test
