+++ keel-core-test/src/main/kotlin/com/netflix/spinnaker/keel/persistence/PeriodicallyCheckedRepositoryTests.kt
@@ -1,6 +1,6 @@
 package com.netflix.spinnaker.keel.persistence
 
-import com.netflix.spinnaker.keel.time.MutableClock
+import com.netflix.spinnaker.time.MutableClock
 import dev.minutest.junit.JUnit5Minutests
 import dev.minutest.rootContext
 import java.time.Clock
