+++ keel-test/keel-test.gradle.kts
@@ -9,8 +9,5 @@ dependencies {
   api(project(":keel-igor"))
   api(project(":keel-clouddriver"))
   implementation("io.mockk:mockk")
-  implementation("org.junit.jupiter:junit-jupiter-api")
-  implementation("io.strikt:strikt-core")
-  implementation("dev.minutest:minutest")
   implementation(project(":keel-spring-test-support"))
 }
